split out fake platformdataapi
