+++ fake_dependencies/platform-data-api/FakePlatformDataApi.App.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FakePlatformDataApi.App", "FakePlatformDataApi.App\FakePlatformDataApi.App.csproj", "{DE00B545-C8A2-4447-9B6C-BE5F91E79C4B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{DE00B545-C8A2-4447-9B6C-BE5F91E79C4B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DE00B545-C8A2-4447-9B6C-BE5F91E79C4B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DE00B545-C8A2-4447-9B6C-BE5F91E79C4B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DE00B545-C8A2-4447-9B6C-BE5F91E79C4B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
