+++ src/SelfService.Tests/Comparers/CapabilityComparer.cs
@@ -38,7 +38,16 @@ public class CapabilityComparer : IEqualityComparer<Capability?>
 
     public int GetHashCode(Capability obj)
     {
-        return HashCode.Combine(obj.Id, obj.Name, obj.Description, obj.Status, obj.CreatedAt, obj.ModifiedAt, obj.CreatedBy, obj.ModifiedBy);
+        return HashCode.Combine(
+            obj.Id,
+            obj.Name,
+            obj.Description,
+            obj.Status,
+            obj.CreatedAt,
+            obj.ModifiedAt,
+            obj.CreatedBy,
+            obj.ModifiedBy
+        );
     }
 }
