+++ src/SelfService/Properties/launchSettings.json
@@ -16,7 +16,7 @@
         "SS_TOPDESK_API_GATEWAY_API_KEY": "dummy",
         "SS_PROMETHEUS_API_ENDPOINT": "http://localhost:9090",
         "SS_PLATFORM_DATA_API_KEY": "dummykey",
-        "SS_PLATFORM_DATA_ENDPOINT": "http://localhost:5051"
+        "SS_PLATFORM_DATA_ENDPOINT": "http://localhost:8070"
       }
     }
   }
